requirements.txt
pandas == 1.5.3
streamlit == 1.21.0
selenium == 4.13.0
openpyxl == 3.1.0
Pillow == 9.4.0
